requirements.txt
ConfigArgParse==1.1
matplotlib==3.2.0
opencv-python==4.3.0.38
pandas==1.0.2
PyYAML==5.3
scikit-image==0.16.2
seaborn==0.10.0
tensorflow==2.5.0
tqdm==4.43.0
pytorch